Added JSTL libraries which we may need when we get into page design. This is generally the best option for including dynamic html (different things displayed based on conditions) rather than having java code with tons of printouts.
All jsp pages now have an associated java helper file. Code will be kept
in the java files when possible to keep the jsp files clean.

Separated code for converting google auth code to auth token from the
sample calendar method into its own method in GoogleAuthHelper.

When google auth code is returned, it is now automatically converted to
a Credential token and stored in the session. This can be used for all
http requests from google without having to convert the code to a token
every time (google will still verify the token before returning the data
we want).