Added slight validation for ics feed urls, and feedback when url fails that validation.
New submit button on feed selection page to go to feed results page,
which is disabled when no feeds are saved.

Added empty feed results page that will later show the user what events
were read as inputs, which will be sent to the output, which will not be
sent (duplicates), and a button to do the output transfer.